Move UserController#change_password to UsersController#change_password
- We adapted the routes, now the change_password is accessible under an
user i.e /users/$USER/change_password
- Adapted the change password dialog to use the new route
- The method change_password requires login
- Added more specs to cover other scenarios
