+++ src/api/app/controllers/webui/user_controller.rb
@@ -6,31 +6,4 @@ class Webui::UserController < Webui::WebuiController
       redirect_to user_path(User.possibly_nobody)
     end
   end
-
-  def change_password
-    user = User.session!
-
-    unless @configuration.passwords_changable?(user)
-      flash[:error] = "You're not authorized to change your password."
-      redirect_back fallback_location: root_path
-      return
-    end
-
-    if user.authenticate(params[:password])
-      user.password = params[:new_password]
-      user.password_confirmation = params[:repeat_password]
-
-      if user.save
-        flash[:success] = 'Your password has been changed successfully.'
-        redirect_to action: :show, user: user
-      else
-        flash[:error] = "The password could not be changed. #{user.errors.full_messages.to_sentence}"
-        redirect_back fallback_location: root_path
-      end
-    else
-      flash[:error] = 'The value of current password does not match your current password. Please enter the password and try again.'
-      redirect_back fallback_location: root_path
-      return
-    end
-  end
 end
