+++ src/api/app/controllers/webui/users_controller.rb
@@ -1,5 +1,5 @@
 class Webui::UsersController < Webui::WebuiController
-  before_action :require_login, only: [:index, :edit, :destroy, :update]
+  before_action :require_login, only: [:index, :edit, :destroy, :update, :change_password]
   before_action :require_admin, only: [:index, :edit, :destroy]
   before_action :check_displayed_user, only: [:show, :edit, :update]
 
@@ -105,6 +105,33 @@ class Webui::UsersController < Webui::WebuiController
 
   def tokens
     render json: User.autocomplete_token(params[:q])
+  end
+
+  def change_password
+    user = User.session!
+
+    unless @configuration.passwords_changable?(user)
+      flash[:error] = "You're not authorized to change your password."
+      redirect_back fallback_location: root_path
+      return
+    end
+
+    if user.authenticate(params[:password])
+      user.password = params[:new_password]
+      user.password_confirmation = params[:repeat_password]
+
+      if user.save
+        flash[:success] = 'Your password has been changed successfully.'
+        redirect_to action: :show, login: user
+      else
+        flash[:error] = "The password could not be changed. #{user.errors.full_messages.to_sentence}"
+        redirect_back fallback_location: root_path
+      end
+    else
+      flash[:error] = 'The value of current password does not match your current password. Please enter the password and try again.'
+      redirect_back fallback_location: root_path
+      return
+    end
   end
 
   private
