+++ src/api/app/views/webui/user/_password_dialog.html.haml
@@ -3,7 +3,7 @@
     .modal-content
       .modal-header
         %h5.modal-title#branch-modal-label Change Your Password
-      = form_tag(user_change_password_path) do
+      = form_tag(change_password_user_path(User.possibly_nobody)) do
         .modal-body
           .form-group
             = label_tag :password, 'Current Password:'
